Move MOM_generic_tracer and make a stub
MOM_generic_tracer to the GFDL_ocean_BGC directory to allow more
felxebile interfaces between MOM6 and ocean_BGC.

MOM_generic_tracer can be converted to a stub version, currently
this was done with comments, but eventually that code can be removed.

Users of the generic tracers package should move a version of
MOM_generic_tracer previous to this commit to their ocean_BGC if using
a version of MOM6 with this commit, but a version of ocean_BGC that
does not include MOM_generic_tracer.

Contact:
Theresa.Morrison@noaa.gov
Yi-Cheng.Teng@noaa.gov